wip: move things around
